# feat: Implement dynamic benchmark summary with real performance metrics